fix: ensure semanticdb version is pulled from scalafix
Since you're using `Global / semanticdbEnabled := true` this tells sbt
to automatically include the semanticdb compiler plugin. This can be out
of date, even in the newest sbt versions. You can see the version by
doing an `sbt semanticdbVersion`. Without the change I made here it will
default to 4.4.28, which isn't published for 2.13.8, which is why the
build failed before.

This change adds in a small fix to instead pull the `semanticdbVersion`
from the Scalafix plugin which is used, which is why you have this
setting in the first place. This will ensure that as a new Scala version
comes out, you bump scalafix, which will also bump your
semanticdbVersion.